paystack and flutterwave helpers

+++ routes/api.php
@@ -3,15 +3,17 @@
 use \Illuminate\Support\Facades\Route;
 use \RaadaaPartners\RaadaaBase\Http\Controllers\ResourceController;
 
-$prefix = config('endpoints.prefix')? config('endpoints.prefix') : 'general';
+if (config('raadaa.set_routes')) {
+    $prefix = config('endpoints.prefix')? config('endpoints.prefix') : 'general';
 
-Route::group(['as' => 'raadaa.'], function () use($prefix){
-    //resource controller routes
-    Route::as('resources.')->prefix($prefix)->group(function () {
-        Route::get('{endpoint}', [ResourceController::class, 'index'])->name('index');
-        Route::post('{endpoint}', [ResourceController::class, 'store'])->name('store');
-        Route::get('{endpoint}/{id}', [ResourceController::class, 'show'])->name('show');
-        Route::match(['POST', 'PATCH', 'PUT'],'{endpoint}/{id}', [ResourceController::class, 'update'])->name('update');
-        Route::delete('{endpoint}/{id}', [ResourceController::class, 'destroy'])->name('delete');
+    Route::group(['as' => 'raadaa.'], function () use($prefix){
+        //resource controller routes
+        Route::as('resources.')->prefix($prefix)->group(function () {
+            Route::get('{endpoint}', [ResourceController::class, 'index'])->name('index');
+            Route::post('{endpoint}', [ResourceController::class, 'store'])->name('store');
+            Route::get('{endpoint}/{id}', [ResourceController::class, 'show'])->name('show');
+            Route::match(['POST', 'PATCH', 'PUT'],'{endpoint}/{id}', [ResourceController::class, 'update'])->name('update');
+            Route::delete('{endpoint}/{id}', [ResourceController::class, 'destroy'])->name('delete');
+        });
     });
-});
+}
